requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
openai>=1.3.0
typing-extensions>=4.7.0
python-dotenv>=1.0.0
requests>=2.31.0 